Move `createNode` to react flow util
